+++ CHANGELOG.md
@@ -5,17 +5,54 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/)
 principles.
 
+## [1.4.1]
+
+### Added
+
+- `InvoiceOutput` type. Thanks @Johnny99211!
+
+## [1.4.0]
+
+### Added
+
+- support for PHP 8
+
+## [1.3.3]
+
+### Fixed
+
+- removed duplicate array key in `TypeFactory`
+
+## [1.3.2]
+
+### Added
+
+- `AddressGroups` and `AddressGroupsGet` types. Thanks @Johnny99211!
+
+## [1.3.1]
+
+### Added
+
+- `Address` type. Thanks @Johnny99211!
+
 ## [1.3.0]
+
+### Added
+
+- `MultiRequest` class for easily sending multiple records to the Collmex API. Thanks @timopaul
+- `AbstractRequest` and `RequestInterface`
+- new types `PRDGRP` and `PRODUCT_GROUPS_GET`
 
 ### Changed
 
 - `TypeFactory` now uses a mapping array instead of a huge switch/case for creating instances of response records.
+- `Request` extends the new `AbstractRequest` class
 
 ## [1.2.1]
 
 ### Added
 
-- `Voucher` and `VoucerGet` types. Thanks @flowdee!
+- `Voucher` and `VoucherGet` types. Thanks @flowdee!
 
 ## [1.2.0]